Refine sub-cell division trees based on lineage barcodes and cell states

Refine sub-cell division trees based on lineage barcodes and cell states

Usage

subtree_refinement(
  Trees_initial,
  state_lineages,
  barcodes_lineages,
  N_char,
  state_labels_lineages,
  lambda1,
  lambda2,
  maxIter = 100,
  repeat_time = 10
)

Arguments

Trees_initial

List of initial sub-cell division trees (by state lineage).

state_lineages

List of state lineage paths.

barcodes_lineages

List of barcode matrices per state lineage.

N_char

Total number of sites.

state_labels_lineages

List of state labels per state lineage.

lambda1

Weight for cell state score.

lambda2

Weight for barcode score.

maxIter

Maximum number of iterations.

repeat_time

Number of iterations required for convergence.

Value

A list containing:

bestsubtree

List of refined sub-cell division trees (one per state lineage).

bestsubtreescore

List of corresponding best scores for each refined tree.

total_time

Total running time.